==> 3_fit_knn.R <==
# Define and fit ...

# load packages ----
library(tidyverse)
library(tidymodels)
library(here)

# handle common conflicts
tidymodels_prefer()

# load training data

# load pre-processing/feature engineering/recipe

# model specifications 

# define workflows

# hyperparameter tuning values 

# fit workflows/models

# write out results (fitted/trained workflows)
